Renamed package name and applicationID to: com.lokico.PSWind
