+++ .idea/.name
@@ -1,1 +1,1 @@
-PWWindMatDesAct
+PSWind
