+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.lokico.pwwindmatdesact">
+    package="com.lokico.PSWind">
 
     <application
         android:allowBackup="true"
