+++ app/src/main/java/com/lokico/PSWind/OmniMap.java
@@ -1,12 +1,7 @@
-package com.lokico.pwwindmatdesact;
+package com.lokico.PSWind;
 
-import android.annotation.SuppressLint;
-import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.os.Handler;
-import android.view.MotionEvent;
-import android.view.View;
 
 /**
  * An example full-screen activity that shows and hides the system UI (i.e.
