+++ app/src/main/java/com/lokico/PSWind/TJsNorthSound.java
@@ -1,11 +1,7 @@
-package com.lokico.pwwindmatdesact;
+package com.lokico.PSWind;
 
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
 
 public class TJsNorthSound extends AppCompatActivity {
     @Override
