+++ app/src/main/res/layout/activity_noa.xml
@@ -8,6 +8,6 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:background="@drawable/noaa"
-    tools:context="com.lokico.pwwindmatdesact.NOAA">
+    tools:context="com.lokico.PSWind.NOAA">
 
 </RelativeLayout>
