+++ app/src/main/res/layout/activity_omni_map.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/map"
-    tools:context="com.lokico.pwwindmatdesact.OmniMap">
+    tools:context="com.lokico.PSWind.OmniMap">
 
     <!-- The primary full-screen view. This can be replaced with whatever view
          is needed to present your content, e.g. VideoView, SurfaceView,
