+++ app/src/main/res/layout/activity_tjs_north_sound.xml
@@ -8,5 +8,5 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:background="@drawable/tj_north_soundpng"
-    tools:context="com.lokico.pwwindmatdesact.TJsNorthSound">
+    tools:context="com.lokico.PSWind.TJsNorthSound">
 </RelativeLayout>
