+++ app/src/main/res/layout/activity_tjs_seattle.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/tj_seattle"
-    tools:context="com.lokico.pwwindmatdesact.TJsSeattle">
+    tools:context="com.lokico.PSWind.TJsSeattle">
 
     <!-- The primary full-screen view. This can be replaced with whatever view
          is needed to present your content, e.g. VideoView, SurfaceView,
