+++ app/src/main/res/layout/content_main.xml
@@ -11,7 +11,7 @@ xmlns:android="http://schemas.android.com/apk/res/android"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.lokico.pwwindmatdesact.MainActivity"
+    tools:context="com.lokico.PSWind.MainActivity"
     tools:showIn="@layout/activity_main"
     android:background="@drawable/backgroudimage">
